Refactor some stuff into new module
