+++ tests/resources/test_buffer.py
@@ -124,31 +124,31 @@ def test_chunk_size_small():
     # Small buffers have just one chunk
     a = np.zeros((10, 1), np.uint8)
     b = gfx.Buffer(a)
-    assert b._chunk_itemsize == 10
+    assert b._chunk_size == 10
     assert b._chunk_map.size == 1
 
     # Goes up to 256 bytes
     a = np.zeros((256, 1), np.uint8)
     b = gfx.Buffer(a)
-    assert b._chunk_itemsize == 256
+    assert b._chunk_size == 256
     assert b._chunk_map.size == 1
 
     # Beyond 256 bytes, chunking kicks in
     a = np.zeros((257, 1), np.uint8)
     b = gfx.Buffer(a)
-    assert b._chunk_itemsize == 256
+    assert b._chunk_size == 256
     assert b._chunk_map.size == 2
 
     # Happy scaling to 16 pieces
     a = np.zeros((16 * 256, 1), np.uint8)
     b = gfx.Buffer(a)
-    assert b._chunk_itemsize == 256
+    assert b._chunk_size == 256
     assert b._chunk_map.size == 16
 
     # Stays about 16 pieces, but changes itemsize
     a = np.zeros((16 * 256 + 1, 1), np.uint8)
     b = gfx.Buffer(a)
-    assert b._chunk_itemsize == 257
+    assert b._chunk_size == 257
     assert b._chunk_map.size == 16
 
     # --- 4 x float -> itemsize is 16 bytes
@@ -156,31 +156,31 @@ def test_chunk_size_small():
     # Small buffers have just one chunk
     a = np.zeros((10, 4), np.float32)
     b = gfx.Buffer(a)
-    assert b._chunk_itemsize == 10
+    assert b._chunk_size == 10
     assert b._chunk_map.size == 1
 
     # Goes up to 256 bytes
     a = np.zeros((16, 4), np.float32)
     b = gfx.Buffer(a)
-    assert b._chunk_itemsize == 16
+    assert b._chunk_size == 16
     assert b._chunk_map.size == 1
 
     # Beyond 256 bytes, chunking kicks in
     a = np.zeros((17, 4), np.float32)
     b = gfx.Buffer(a)
-    assert b._chunk_itemsize == 16
+    assert b._chunk_size == 16
     assert b._chunk_map.size == 2
 
     # Happy scaling to 16 pieces
     a = np.zeros((256, 4), np.float32)
     b = gfx.Buffer(a)
-    assert b._chunk_itemsize == 16
+    assert b._chunk_size == 16
     assert b._chunk_map.size == 16
 
     # Stays about 16 pieces, but changes itemsize
     a = np.zeros((257, 4), np.float32)
     b = gfx.Buffer(a)
-    assert b._chunk_itemsize == 17
+    assert b._chunk_size == 17
     assert b._chunk_map.size == 16
 
 
@@ -190,50 +190,50 @@ def test_chunk_size_large():
 
     a = np.zeros((50 * 2**20, 1), np.uint8)
     b = gfx.Buffer(a)
-    assert b._chunk_itemsize == 2**20
+    assert b._chunk_size == 2**20
 
     a = np.zeros((200 * 2**20, 1), np.uint8)
     b = gfx.Buffer(a)
-    assert b._chunk_itemsize == 2**20
+    assert b._chunk_size == 2**20
 
     a = np.zeros((50 * 65536, 4), np.float32)
     b = gfx.Buffer(a)
-    assert b._chunk_itemsize == 2**16  # 2**16 * 16 == 2**20
+    assert b._chunk_size == 2**16  # 2**16 * 16 == 2**20
 
     a = np.zeros((200 * 65536, 4), np.float32)
     b = gfx.Buffer(a)
-    assert b._chunk_itemsize == 2**16
+    assert b._chunk_size == 2**16
 
 
 def test_custom_chunk_size():
     # Custom chunk size can get it lower (one row is 16 bytes)
     a = np.zeros((16, 4), np.float32)
     b = gfx.Buffer(a, chunksize=16)
-    assert b._chunk_itemsize == 1
+    assert b._chunk_size == 1
     assert b._chunk_map.size == 16
 
     # Custom chunk size rounds to item size
     a = np.zeros((16, 4), np.float32)
     b = gfx.Buffer(a, chunksize=10)
-    assert b._chunk_itemsize == 1
+    assert b._chunk_size == 1
     assert b._chunk_map.size == 16
 
     # Custom chunk size rounds to item size
     a = np.zeros((26, 4), np.float32)
     b = gfx.Buffer(a, chunksize=1000)
-    assert b._chunk_itemsize == 26
+    assert b._chunk_size == 26
     assert b._chunk_map.size == 1
 
     # Custom chunk size caps to 1 byte
     a = np.zeros((200, 1), np.uint8)
     b = gfx.Buffer(a, chunksize=-1)
-    assert b._chunk_itemsize == 1
+    assert b._chunk_size == 1
     assert b._chunk_map.size == 200
 
     # Custom chunk size caps to 1 element
     a = np.zeros((200, 4), np.float32)
     b = gfx.Buffer(a, chunksize=-1)
-    assert b._chunk_itemsize == 1
+    assert b._chunk_size == 1
     assert b._chunk_map.size == 200
